+++ packages/photo-companion/src/lib/components/navigation/navigation.menu.tsx
@@ -1,5 +1,8 @@
-import { For, type ParentProps } from "solid-js";
+import { For } from "solid-js";
+import { Button, IconClose, Link } from "ui";
 
+import { ROUTE_CHANGELOG, VERSION } from "@lib/consts";
+import { useTranslation } from "@lib/context/translation";
 import { useSupportsBridges, useSupportsLights } from "@lib/hooks";
 
 import { NAVIGATION_TAB_DATA } from "./navigation.const";
@@ -10,6 +13,10 @@ import styles from "./navigation.menu.module.css";
 const classes = {
 	link: styles.link
 };
+
+interface Props {
+	onClose: VoidFunction;
+}
 
 function useNavItemGroups() {
 	const supportsLights = useSupportsLights();
@@ -32,25 +39,51 @@ function useNavItemGroups() {
 	]);
 }
 
-export function NavigationMenu(props: ParentProps) {
+export function NavigationMenu(props: Props) {
+	const { t } = useTranslation();
 	const groups = useNavItemGroups();
 
 	return (
-		<nav class={styles.menu}>
-			<For each={groups()}>
-				{group => (
-					<section class={styles.section}>
-						<For each={group}>
-							{item => (
-								<NavigationItem classes={classes} {...item} />
-							)}
-						</For>
-					</section>
-				)}
-			</For>
-			<form class={styles.form}>
-				{props.children}
-			</form>
-		</nav>
+		<>
+			<header class={styles.header}>
+				<h2>{t().LABEL.MENU}</h2>
+				<Button icon onClick={props.onClose} type="button">
+					<IconClose />
+				</Button>
+			</header>
+			<nav class={styles.menu}>
+				<For each={groups()}>
+					{group => (
+						<section class={styles.section}>
+							<For each={group}>
+								{item => (
+									<NavigationItem classes={classes} {...item} />
+								)}
+							</For>
+						</section>
+					)}
+				</For>
+			</nav>
+			<footer class={styles.footer}>
+				<dl class={styles.info}>
+					<div>
+						<dt>Version:</dt>
+						<dd>
+							<Link href={ROUTE_CHANGELOG}>
+								v.{VERSION}
+							</Link>
+						</dd>
+					</div>
+					<div>
+						<dt>Commit:</dt>
+						<dd>
+							<Link href="https://github.com/ericrovell/photo-companion/commit/__COMMIT_HASH__">
+								#__COMMIT_HASH__
+							</Link>
+						</dd>
+					</div>
+				</dl>
+			</footer>
+		</>
 	);
 }